test: move the WorkloadRun timeout cases to TestCaseParser
Follows pkg/catalog/gpu_defaults_test.go, as pkg/controller does
throughout.

Four cases under testdata/workloadrun-timeout, one per way the timeout
stayed nil: no orchestration block at all, an orchestration block with no
timeout, an unparseable value whose parse error was discarded, and the
user's own value winning.

The isJobTimedOut check stayed a plain Go test. It is a single-value
assertion, and it is the one that proves the bound is wired through rather
than merely stored.
