+++ pkg/controller/testdata/workloadrun-timeout/orchestration-without-timeout/expected.json
@@ -1,0 +1,3 @@
+{
+  "timeoutPerJob": "24h0m0s"
+}
