+++ pkg/controller/testdata/workloadrun-timeout/orchestration-without-timeout/input.yaml
@@ -1,0 +1,1 @@
+orchestration: {}
