+++ pkg/controller/testdata/workloadrun-timeout/users-own-value/expected.json
@@ -1,0 +1,3 @@
+{
+  "timeoutPerJob": "45m0s"
+}
