+++ pkg/controller/testdata/workloadrun-timeout/users-own-value/input.yaml
@@ -1,0 +1,2 @@
+orchestration:
+  timeoutPerJob: 45m
